Added dt_min and dt_max parameters to Image class
The clamping on dT is now applied to every image (even constant images) but the range can be better controlled.

Also fixed a couple of failing tests.  The failing NDVI test was a result of Red 0.2 and NIR 0.3 not computing exactly to an NDVI of 0.2.  The failing dT was a result of not setting the DOY on the Image class.
Still need to implment elr_flag.