Cleaned up Tcorr export code
Add code for Landsat 4 (but didn't uncomment).

Removed study area (export_geom) filtering.

Changed daily code to start with Tmax image and then add daily Tcorr.  The benefit of this is that each day will be exported even if it has no data, so that the scripts don't keep trying to re-export missing days.  Changed the monthly to be similar so that code starts with Tmax image set to the default Tcorr value and then updates if there is data (over the scene count minimum).

Added list of SPACECRAFT_IDs used for daily.  This requires a couple extra "If" calls which it would be nice to remove at some point.