Miscellaneous updates for getting asset generating working
Added "custom" interpolation time interval that will interpolate the full date range (but return an ImageCollection with a single image).  Made this the default t_interval on the interpolation method.

Implemented a collection filter_args approach for applying arbitrary filters to the collections.  For now, only an "equals" metdata filter is working.

Added a "mask" parameter to the Image class.  Something like this is needed for getting at the interpolation/overpass count output.

elr_flag can be a bool or a string representation of a bool.