Testing only aggregating count/mask images to new daily collection
Since the interpolation is using the time band which is already 0 UTC, I don't think the interpolated variables need to be aggregated to a new collection before interpolating.  The aggregation is needed to get the correct count value, but hopefully this will simplify the calculation for the other variables.

Changed etr_source and etr_band variables to not have defaults and they can be passed in class init, model_args, or interpolate call.