Lag Tcorr daily image export by 4 days
Landsat images currently have a processing lag of ~3 days but building Tcorr immediately on the 3rd day causes it to be only partially complete.  Lagging by 4 days guarantees that most if not all of the images that are going to be available are available.  This may need to be updated if the Landsat ingest speeds up at some point.