Added tcorr daily image cleanup script to remove duplicate Tcorr daily images
The Tcorr daily image script is setup to export a Tcorr image everytime it is run, even if one already exists for that day.  Each image has the export date in the image ID and the SSEBop code looks for the Tcorr image with the greatest export date.  This ensures that a Tcorr image is always available even if we update Tcorr.

Also, switched to set literals in the image test to match collection tests.