Updated path/row logging in Tcorr images
The original list of all WRS2 path/rows was getting to be too long for the max property sizes.  I changed it so the path/rows are saved as a dictionary of lists where the key is the path and the row is in the list.

Also added code to only export a subset of the DAYMET image.  This will probably need to be revisited.