Commented out Tcorr FEATURE and IMAGE options and changed how Tcorr index was used
It made more sense for the SCENE Tcorr to return the index as a property of the image instead of as a separate item.  This is also much easier to code sense the Tcorr images already have the index property so the function just has to return the image that was used and the property will already be set.

Bumped version number.