Modified Tcorr scene exports to remove transparency masks
The transparency masks were being set to the Tcorr value since the updateMask call was being made on the unmasked Tcorr image (i.e. the Tcorr value of 0.98 was being used as the transparency mask).  Now the updateMask call is made with the raw mask image that is either 0 or 1.