Variable renaming and formatting
Switched all the references of "cfact" to "tcorr" to be consistent with our other code and methods.  We could fully switch over to "cfact" at some point if that would be more consistent with the papers and other documentation though.

Renamed all "5km" variables to "coarse" since I don't think we want to have the grid size as part of the variable names in case we change the grid size at some point.  I switched everything to "coarse" but maybe something like "grid" might be better.