Combined separate tcorr gridded functions
The weights are applied in the main tcorr gridded function based on a flag instead of having separate functions.  The final smoothing is also controlled with a flag.  I still want to make te

Commenting out all of the compositing in the main tcorr function until I can write some test for that part.