Bug fix for tcorr_stats "tcorr_p5" band naming
Switched it to "tcorr_value", although we may want to revisit this and call it something else.

Alos, I had to switch all the test values also.  At some point we might want to make the Tcorr percentile a parameter instead of hardcoding it.  I left the old values in just in case.