Added support for setting the Tmax source as a collection ID
Made a copy of the "DAYMET_MEDIAN_V2" collection called "projects/usgs-ssebop/tmax/daymet_v3_median_1980_2018" and set this as the default until the Daymet v4 collection is ready.

Removed fall back code from the "day of" Tmax calculation since we rarely use this option and instead use the median.  Now the code will just fail if a Tmax image doesn't exist (like for 2020 with Daymet or before 2003 with CIMIS) which is probably better than not being sure which Tmax was used.  If the user wanted to generate these dates then they would need to explicitly switch to the median collection.