Getting "CONUS" feature from an asset instead of building dynamically
This should be quite a bit faster than the existing code that was building up a CONUS asset from all the state features.

I also bumped the tool versions up to the model version (0.1.7) since they are pretty linked at this point and the latest version of the code is needed for the tcorr indexing to work properly.

I'll need to do a release to pypi to make this work in the cloud function.