Added use_joins parameter to Collection.interpolate function
This parameter is passed through to the openet.core.interpolate.daily() function.  It might make more sense to try and read this from kwargs but this seemed a little more explicit/obvious.

Also added a conditional on the compute_et calculation so that it is only applied if et or et_fraction is in the variables list.  This is needed so that the model can be used to build interpolated NDVI images.