Removing deprecated dt and tmax source keyword support
The dt clamping parameters (dt_min and dt_max) can probably be removed since dt clamping is not applied for collection ID sources.

The "reflectance_type" parameter could probably be removed since TOA collections are no longer supported but leaving for now in case it is ever added back in.

Should probably also remove the DYNAMIC tcorr option but want to find a way to keep it if possible.  Need to decide on if support for the tcorr indices should be kept.