LST source updates
Started adding support for falling back on the input LST image if the LST source image does not exist, but ended up not implementing this (for now).  If the user is expecting to use the LST source, it seems better to fail instead of returning an estimate based on the input LST, but that is not ideal when mapping over a collection.  We might want to instead set the LST to a masked image or something so nothing is computed but an image is generated.