Override NDVI and NDWI values for pixels with very low (or negative) reflectance
The build in normalizedDifference function returns no data if the input values are negative, but the reflectance values over water are often negative in collection 2.  Also the normalized difference calculation is very sensitive to small changes when the reflectance values are very low and may not be representative of real differences.

Setting NDWI to zero for pixels with very low reflectance, and setting NDVI to -0.1 for pixels with very low reflectance that are flagged as water in the QA_PIXEL band, and 0 for all other (low reflectance) pixels.  These values, or the thresholds, may need to be changed or adjusted in the future.