75 layered t cold (#76)
* version bump to v0p6p0, new lazy prop for 'ag_landcover', edits to tcorr and added a lc_source property for landcover. set dt_source to v7...

* ag_landcover lazy prop is finished.

* tcorr_FANO is completely rewritten in the style of layered TCold

* focalmax_rad=5 default

* Small fix for type in NALCMS land cover list

Value 14 was repeated twice, but I think the second one should be 15 for croplands.

Also started adding code to structure this section as remaps to make it easier to support other datasets.

* Removing NDWI from tests

* Reformatting of Tcorr code

* Updated lc_source and ag_landcover testing

Add support for setting lc_source as a number for testing.

We might want to rename property to something like tcorr_ag_mask or tcorr_ag_landcover to make it clear what the landcover is for.

* Cleaning up tcorr_not_water_mask function

* Update image.py

* tcorr_not_water_mask attempt using reduceNeighborhood. A few changes to tcorr_FANO lazy prop to make things work as intended.

version bump to make verification of testing easier.

* New not_water_mask()

version bump to make verification of testing easier.

* Removed buffering from not water mask creation

Started adding NDWI back in for defining not water

Flipped the bands so that NDWI is computed as SWIR - Green instead of Green - SWIR so that negative values correspond to water.

Still need to update the not water mask tests.

* Testing out removing extra reproject calls

And small bug fix to add NDWI back into building the image class in from_image_id.

* Adjustments to gridsizes and reduced coarse smoothing window from 5 pixels to 1

* Added support for using setting lc_source to NLCD

The parameter can be the annual NLCD image or collection ID, or the 2019 and 2021 releases.  The code will select the closest image in time if the collection ID is used.

projects/sat-io/open-datasets/USGS/ANNUAL_NLCD/LANDCOVER
projects/sat-io/open-datasets/USGS/ANNUAL_NLCD/LANDCOVER/Annual_NLCD_LndCov_2023_CU_C1V0
USGS/NLCD_RELEASES/2021_REL/NLCD
USGS/NLCD_RELEASES/2021_REL/NLCD/2021

* Bumping minimum Python and GEE versions

* Apply mask from input image to LST source images

The sharpened LST images are not masked for clouds, so the cloud mask from the source image needs to be applied to them.

* Update image.py

Use the NDVI image to get the cloud mask since the LST image will have the emissivity holes in it

* commited smooth_mixed_landscape_tcorr_ag_plus_veg commented out. Already not used. Any savings?

* commited smooth_mixed_landscape_tcorr_ag_plus_veg commented out.version bump to track savings

* commiting

* No longer smoothing NDVI everywhere...

* G Senay's latest flow chart.

* G Senay's latest flow chart with one additional smoothing from Tc Veg 4800m into Tc Veg 120m to remove artifacts...

* doing another smoothing of coarse ag pixels to increase coverage...

* v 0.6.9, Only smoothing tcorr at the end, once, and making a tweat to water getting it's lst

* v0.7.1 toml fixed for windows recursive build directories... added a few steps to lazy props to make visualization easier...

* v0.7.5 5km now uses the 120 meters AVERAGE NDVI rather than the INDIVIDUAL NDVI pixels..

* v0.7.5 5km now uses the 120 meters AVERAGE NDVI rather than the INDIVIDUAL NDVI pixels.. 5km only smoothed once...
Some lazy properties added to more easily visualize in a NB for testing.

* Changing the final projection of the Tc to allow for edge cases to be handled more gracefully.

* Threshold set from 0.35 to 0.4. The difference was seen in the NW corner of this scene LANDSAT/LC09/C02/T1_L2/LC09_028032_20240524. 0.4 is an NDVI cutoff in the lit. that is defensible.

* halving the m_pixels_coars value to see if we get any savings...

* Committing a good version that fixes the winter slop. Doing partial conus run with it.

* Best version yet

* Updating Tcorr not water mask tests

* Update test_c_image.py

Burning in current test value

* Couple of small fixes and updates

* version 0.6.0 still using tmax to make tcorr (vestigial) tests are passing, values burnt in.

* cleaning up tcorr_fano lazy prop

* passing tests.

---------

Co-authored-by: Charles Morton <cgmorton@users.noreply.github.com>